Automated TiledElevationCoverage LevelSet configuration (#654)
The LevelSet parameters required during TiledElevationCoverage construction are much better as an implementation detail. The caller has already supplied the minimum necessary information needed to configure a LevelSet: sector and resolution. Requiring the caller to specify a level count is unnecessary and potentially dangerous. Level count is trivial to derive from resolution, but its easy for an uniformed caller to specify a level count that does not match the data. Additionally, the level zero delta and tile width/height are carefully chosen to balance operation and performance, and should not be specified by the caller. The level zero delta, tile width, and tile height are set to constant values known to work well with TiledElevationCoverage. The level count is derived from the resolution. The resultant LevelSet has a last level that either matches or exceeds the specified coverage resolution.

Note: The LevelSet defaults are left as inline constants, as TiledElevationCoverage's usage of LevelSet is temporary.

Relates to #486