Modified how ElevationCoverage.elevationsForGrid interprets resolution (#656)
* Modified how ElevationCoverage.elevationsForGrid interprets resolution

The function elevationsForGrid derived from the original ElevationModel, which used targetResolution to select a level in the model's tile pyramid. Since implementing composable elevations (#575), this function appears in both ElevationModel and ElevationCoverage, but serves a different purpose in each. Given that, the argument list need not be the same, and in fact should not be the same. ElevationModel needs targetResolution to let the caller influence which coverage is chosen to fulfil the request. ElevationCoverage does not need targetResolution. Allowing the caller to specify a resolution to ElevationCoverage is unnecessary and potentially dangerous. The grid itself provides an implicit resolution the coverage can use to select a level, and letting the caller override that resolution risks the caller accessing more data than can fit in memory.

Relates to #422

* Adjusted ElevationModel unit tests to reflect ElevationCoverage changes